Add Visual.Renderer module
