+++ straminata.cabal
@@ -16,7 +16,7 @@ library
   hs-source-dirs:      src
   ghc-options:         -Wall -fno-warn-type-defaults
   exposed-modules:
-    Straminata, WindowConstants, Model.CommonTypes, Util.Common
+    Straminata, WindowConstants, Model.CommonTypes, Util.Common, Visual.Renderer
   build-depends:       base >= 4.7 && < 5
                      , gloss
   default-language:    Haskell2010
